+++ src/components/Tag/index.ts
@@ -1,0 +1,5 @@
+import Tag, { TagProps } from "./Tag";
+import TagList, { TagListProps } from "./TagList";
+
+export { Tag, TagList };
+export type { TagProps, TagListProps };