+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 
 * use option `--outputCss` for `yarn compile-scss` to get the transpiled CSS echoed out
+* `<ConfidenceValue/>` element: combines a value and a bar
 * `<Badge />` element to add more context like icons, text or numbers to another element
 * `<Depiction />` element to include different types of images controlling of resizing, ratio, shape
 * `<Tag />`: add support for `intent` property
@@ -20,6 +21,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 * `<TextField />` and `<TextArea />`:
     * `invisibleCharacterWarning`: callback to warn of invisible, hard to spot characters in the input text.
 * `<SilkActivityControl />`: parameter `executePrioritized` that is executed when the 'start prioritized' button is clicked while an activity is waiting for execution.
+* Carbon based `<Table />`, `<TableExpandHeader />` and `<TableExpandRow />` elements, other table elements are still used directly from the Carbon library.
 
 ### Changed